// frontend/src/api/generated/models/location-full-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AreaType } from './area-type';
// May contain unused imports in some cases
// @ts-ignore
import type { StorageCondition } from './storage-condition';

/**
 * 
 * @export
 * @interface LocationFullDetail
 */
export interface LocationFullDetail {
    /**
     * 
     * @type {number}
     * @memberof LocationFullDetail
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'locationCode': string;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'locationName'?: string;
    /**
     * 
     * @type {number}
     * @memberof LocationFullDetail
     */
    'warehouseId': number;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'warehouseCode': string;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'warehouseName'?: string;
    /**
     * 
     * @type {number}
     * @memberof LocationFullDetail
     */
    'areaId': number;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'areaCode': string;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'areaName'?: string;
    /**
     * 
     * @type {AreaType}
     * @memberof LocationFullDetail
     */
    'areaType': AreaType;
    /**
     * 
     * @type {StorageCondition}
     * @memberof LocationFullDetail
     */
    'storageCondition'?: StorageCondition;
    /**
     * 
     * @type {number}
     * @memberof LocationFullDetail
     */
    'buildingId': number;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'buildingCode': string;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'buildingName'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof LocationFullDetail
     */
    'isActive': boolean;
    /**
     * 
     * @type {number}
     * @memberof LocationFullDetail
     */
    'version': number;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof LocationFullDetail
     */
    'updatedAt': string;
}